move name.go from cluster to p2p
