+++ p2p/name.go
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License along with
 // this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package cluster
+package p2p
 
 import (
 	"crypto/ecdsa"
